Add Android Project (#2)
* Add first Android Project

* Add android pipeline

* Fix artifacts

* Add forms and library on android

* Fix android and fix vote to double

* Remove Android build when build API

* Fix to compile in release

* Fix android package name

* Add camera permission

* Add permission request on Android

* Add privacy

+++ Game.Planning.Poker.Mobile/Domain/GameService.cs
@@ -167,7 +167,7 @@ namespace Game.Planning.Poker.Mobile.Domain
             return Task.CompletedTask;
         }
 
-        private async Task InternalVote(Player player, int vote)
+        private async Task InternalVote(Player player, double vote)
         {
             await this.Vote(player, vote);
